requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
psycopg2-binary
alembic
passlib[argon2]==1.7.4
argon2-cffi
python-jose[cryptography]
python-multipart